requirements.txt
asgiref==3.7.2
astroid==3.0.3
attrs==23.2.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
dill==0.3.8
Django==4.2
django-cors-headers==4.3.1
django-environ==0.11.2
django-filter==23.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
frozendict==2.4.0
google-api-core==2.17.1
google-auth==2.28.2
google-cloud-secret-manager==2.18.3
googleapis-common-protos==1.63.0
grpc-google-iam-v1==0.13.0
grpcio==1.62.1
grpcio-status==1.62.1
gunicorn==19.9.0
idna==3.6
install==1.3.5
isort==5.13.2
mailjet-rest==1.3.4
mccabe==0.7.0
mock==5.1.0
packaging==24.0
pillow==10.2.0
platformdirs==4.2.0
proto-plus==1.23.0
protobuf==4.25.3
psycopg==3.1.18
psycopg2==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyJWT==2.8.0
pylint==3.0.3
pylint-django==2.5.5
pylint-plugin-utils==0.8.2
pytz==2024.1
requests==2.31.0
rsa==4.9
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
six==1.16.0
sqlparse==0.4.4
tomlkit==0.12.3
typing_extensions==4.9.0
tzdata==2023.4
urllib3==1.26.18
webapp2==2.5.2
